Add Clock 2.0 project
This project enhances the 001 Clock a little. It can be used with both UDP and TCP. UDP only works for setting the time because it doesn't have a return value. If setTime had been declared as throwing an exception in the ice value, this operation wouldn't work on UDP.

Glacier2 can also be used (for TCP).